refactor(daemon): extract runner::inflight submodule (#2642)
Move check_in_flight, log_result, maybe_queue_self_prompt, and
status to runner/inflight.rs.
